% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anc_coeff_to_GeneticSimilarityMatrix.R
\name{anc_coeff_to_GeneticSimilarityMatrix}
\alias{anc_coeff_to_GeneticSimilarityMatrix}
\title{calculate genetic similarity from ancestry coefficients.}
\usage{
anc_coeff_to_GeneticSimilarityMatrix(anc_coef)
}
\arguments{
\item{anc_coef}{a matrix of ancestry coefficients with samples by rows. Each column corresponds to an ancestral population.}
}
\description{
calculate genetic similarity from ancestry coefficients.
}
\details{
Since ancestry coefficients can be interpreted as a propotion of genome derived from a specific ancestral population, this function calculate genetic similarity as the probability of a random genome segment of two individuals derived from the same ancestral population.
}
