% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GGoutlieR.R
\name{plot_GGoutlieR}
\alias{plot_GGoutlieR}
\title{Project GGNet on a geographical map}
\usage{
plot_GGoutlieR(
  geo_coord,
  GGNet_adjm,
  anc_coef = NULL,
  pie_color = NULL,
  p_thres = NULL,
  GeoSP_knn_res = NULL,
  GenSP_knn_res = NULL,
  color_res = 10,
  dot_cex = 0.4,
  map_type = c("geographic_knn", "genetic_knn", "both"),
  geo_xlim = c(-180, 180),
  geo_ylim = c(-90, 90),
  plot_xlim = NULL,
  plot_ylim = NULL,
  only_edges_in_xylim = TRUE,
  pie_r_scale = 1,
  red_alpha = 0.8,
  map_resolution = "high",
  show_knn_pie = FALSE,
  which_sample = NULL
)
}
\arguments{
\item{geo_coord}{a two column matrix or data.frame. the first column is longitude and the second one is latitude.}

\item{GGNet_adjm}{the output of `get_GGNet_adjacency_matrix`}

\item{anc_coef}{a matrix of ancestry coefficients with samples ordered by rows. Ancestry coefficients are used to make pie charts on a geographical map. This argument is optional.}

\item{pie_color}{colors of pie charts. colors are automatically assigned if `pie_color = NULL` (which is the default). This argument is optional.}

\item{p_thres}{a value of significant level. only samples with p values less than `p_thres` are mapped on a geographical map if `p_thres` is provided (the default is `NULL`). This argument is optional.}

\item{GeoSP_knn_res}{an output from `detect_outlier_in_GeoSpace`}

\item{GenSP_knn_res}{an output from `detect_outlier_in_GeneticSpace`}

\item{color_res}{the resolution of color scale}

\item{dot_cex}{the size of dots denoting the positions of samples on a geographical map.}

\item{map_type}{the type of plot to draw. it can be `"geographic_knn"`, `"genetic_knn"` and `"both"`.}

\item{geo_xlim}{values controlling longitude boundaries to select outliers to present on a geographical map. the default is `geo_xlim = c(-180,180)`.}

\item{geo_ylim}{values controlling latitude boundaries to select outliers to present on a geographical map. the default is `geo_ylim = c(-90,90)`.}

\item{plot_xlim}{values controlling longitude boundaries of a map.}

\item{plot_ylim}{values controlling latitude boundaries of a map.}

\item{only_edges_in_xylim}{logic. only the edges with starting points within the given `geo_xlim` and `geo_ylim` will display on a geographical map. the default is `TRUE`.}

\item{pie_r_scale}{a scale controlling the radius of pie charts}

\item{red_alpha}{a value controlling the transparency of red lines. the default is 0.8}

\item{map_resolution}{the resolution of the geographical map. See details in the manual of `rworldmap::getMap()`}

\item{show_knn_pie}{logic. If `TRUE`, the ancestry coefficients of K nearest neighbors of significant samples will display on the map. The default is `FALSE`.}

\item{which_sample}{a string vector. If users want to only}
}
\description{
Project GGNet on a geographical map
}
\details{
Red links on the map denote individual pairs that are genetically similar but geographically remote.
The color depth and thickness of red links are proportional to -log10(p) based on the empirical Gamma distribution obtained from `detect_outlier_in_GeneticSpace`.
Blue links on the map denote individual pairs that are genetically different but geographically close.
The color depth and thickness of blue links are proportional to -log10(p) based on the empirical Gamma distribution obtained from `detect_outlier_in_GeoSpace`
}
