% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ggoutlier.R
\name{summary_ggoutlier}
\alias{summary_ggoutlier}
\title{Summarize GGoutlieR results}
\usage{
summary_ggoutlier(ggoutlier_res)
}
\arguments{
\item{ggoutlier_res}{output from the function `ggoutlier`}
}
\value{
a table contains IDs of outliers and p-values.
}
\description{
Get a summary table from the `ggoutlier` output
}
