% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outlier_in_GeoSpace.R
\name{ggoutlier_geoKNN}
\alias{ggoutlier_geoKNN}
\title{Identify samples genetically different from K nearest geographical neighbors (geographical space KNN)}
\usage{
ggoutlier_geoKNN(
  geo_coord,
  gen_coord,
  min_nn_dist = 100,
  k = NULL,
  klim = c(3, 50),
  s = 100,
  plot_dir = ".",
  w_power = 1,
  p_thres = 0.05,
  n = 10^6,
  multi_stages = TRUE,
  maxIter = NULL,
  keep_all_stg_res = FALSE,
  warning_minR2 = 0.9,
  cpu = 1,
  verbose = TRUE
)
}
\arguments{
\item{geo_coord}{a two column matrix or data.frame. the first column is longitude and the second one is latitude.}

\item{gen_coord}{a matrix of "sample's coordinates in a genetic space". Users can provide ancestry coefficients or eigenvectors for calculation. If, for example, ancestry coefficients are given, each column corresponds to an ancestral population. Samples are ordered in rows as in `geo_coord`.}

\item{min_nn_dist}{a minimum distance (in a unit of 1 meter) between a given focal sample with its neighbor. The neighboring samples within this distance will not be selected as KNN of a focal sample. Use this argument if you want to avoid searching KNNs within communities. The default is `NULL`. `min_nn_dist` will be adjusted automatically with the `s` parameter if `min_nn_dist` is given. The default is 100 meters.}

\item{k}{number of the nearest neighbor. the default is `NULL`.}

\item{klim}{if `k = NULL`, an optimal k will be searched between the first and second value of `klim`}

\item{s}{a scalar of geographical distance. The default `s=100` scales the distance to a unit of 0.1 kilometer.}

\item{plot_dir}{the path to save plots}

\item{w_power}{a value controlling the power of distance weight in KNN prediction. For example, if `w_power=2`, the weight of KNN is 1/d^2/sum(1/d^2).}

\item{p_thres}{a significance level}

\item{n}{number of samples to draw from the null distribution (to obtain the range of x axis to make a curve plot of null distribution)}

\item{multi_stages}{logic. a multi-stage test will be performed if is `TRUE` (the default is `TRUE`).}

\item{maxIter}{maximal iteration number of multi-stage KNN test.}

\item{keep_all_stg_res}{logic. results from all iterations of the multi-stage test will be retained if it is`TRUE`. (the default is `FALSE`)}

\item{warning_minR2}{the prediction accuracy of KNN is evaluated as R^2 to assess the violation of isolation-by-distance expectation. if any R^2 is larger than `warning_minR2`, a warning message will be reported at the end of your analysis.}
}
\value{
a list including five items. `statistics` is a `data.frame` consisting of the D_g values, p values and a column of logic values showing if a sample is an outlier or not. `threshold` is a `data.frame` recording the significance threshold. `gamma_parameter` is a vector recording the parameter of the heuristic Gamma distribution. `knn_index` and `knn_name` are a `data.frame` recording the K nearest neighbors of each sample.
}
\description{
Identify samples genetically different from K nearest geographical neighbors (geographical space KNN)
}
\examples{
# example 1 -> using ancestry coefficients:
 geo_coord <- read.table("./data/georef_1661ind_geo_coord_for_locator.txt", header = T, stringsAsFactors = F)
 rownames(geo_coord) <- geo_coord[,1]
 geo_coord <- geo_coord[,-1]
 anc.coef <- t(as.matrix(read.csv("./data/alstructure_Q_hat_1661inds.csv", header = F, stringsAsFactors = F)))
 sus <- ggoutlier_geoKNN(geo_coord = geo_coord, gen_coord = anc.coef,
                                    plot_dir = "./fig", p_thres = 0.05, w_power = 1,
                                    k = NULL, klim = c(3,100), keep_all_stg_res = F)
 example 2 -> using eigenvectors:
 pc <- read.table("data/gbs_georef_1661inds_PCA.eigenvec", stringsAsFactors = F)
 rownames(pc) <- gsub(pc[,2], pattern = "^0_", replacement = "")
 pc <- apply(pc[,-c(1:2)], 2, function(x){scale(x)}) # removing FID and IID and normalizing data
 sus <- ggoutlier_geoKNN(geo_coord = geo_coord, gen_coord = pc[,1:5],
                                       plot_dir = "./fig", p_thres = 0.05, w_power = 1,
                                       k = NULL, klim = c(3,100), keep_all_stg_res = F)
}
